[configuration] extract confguration parser
