+++ spec/fixtures/echo.conf
@@ -8,4 +8,8 @@ server {
         echo_read_request_body;
         echo $request_body;
     }
+
+    location /json {
+        echo "{}";
+    }
 }
